move column mapper to root on test
