+++ jnosql-mapping/jnosql-mapping-column/src/test/java/org/eclipse/jnosql/mapping/column/ColumnMapperDeleteTest.java
@@ -12,7 +12,7 @@
  *
  *   Otavio Santana
  */
-package org.eclipse.jnosql.mapping.column.query;
+package org.eclipse.jnosql.mapping.column;
 
 import jakarta.nosql.column.ColumnDeleteQuery;
 import jakarta.nosql.mapping.column.ColumnQueryMapper;
